requirements.txt
pillow
matplotlib
pandas
ipython
pytest
svg2png
